pyproject.toml
[project]
name = "jegadeesh-titman-93-rep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "statsmodels>=0.14.5",
    "wrds>=3.1.6",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pandas-stubs>=2.3.2.250827",
    "ruff>=0.13.0",
    "types-openpyxl>=3.1.5.20250914",
    "types-python-dateutil>=2.9.0.20250822",
    "types-xlrd>=2.0.0.20250914",
]
